fix: weight, count, setIndex 타입변경

+++ Workout/Models/SetConfiguration.swift
@@ -8,33 +8,33 @@
 import Foundation
 
 class SetConfiguration {
-  var weight: Int
-  var count: Int
+  var weight: Float
+  var count: UInt
   
   init() {
     self.weight = 0
     self.count = 0
   }
   
-  init(weight w: Int) {
+  init(weight w: Float) {
     self.weight = w
     self.count = 0
   }
   
-  init(count c: Int) {
+  init(count c: UInt) {
     self.weight = 0
     self.count = c
   }
   
-  func updateWeight(with w: Int) {
+  func updateWeight(with w: Float) {
     self.weight = w
   }
   
-  func updateCount(with c: Int) {
+  func updateCount(with c: UInt) {
     self.count = c
   }
   
-  func weightTimesCount() -> Int {
-    return self.weight * self.count
+  func weightTimesCount() -> Float {
+    return self.weight * Float(self.count)
   }
 }
